Bremen Big Data Challenge 2023
Welcome to the Bremen Big Data Challenge 2023!

In this year’s task, measurement data from water samples shall be processed. The measurements were carried out by the Alfred-Wegener-Institut near Helgoland (North Sea) and recorded over a period of 54 years (01.01.1962 - 31.12.2015).

The measurement data is structured as a table. For each measurement there is one entry with 8 values given. Since this is real world measurement data, some data points may be missing. Those are marked as empty entries or with "NA". Also, entries could be noted with an additional "?" to indicate that the recorded value is uncertain. The data is entered in English format (decimal places/decimal separators are indicated by a dot "." and thousands separators by a comma ",").

The development file contains the pure measurement data including missing entries and "?" notes over a period from 01.01.1962 - 31.12.2010.

Based on this data, a forecasting model shall be developed to predict values for the evaluation periods 01.01.2004 - 31.12.2004 and 01.01.2011 - 31.12.2013.

The exact days for which predictions are to be made are specified in the file "bbdc_2023_AWI_data_evaluate_skeleton.csv" ("Datum", "Uhrzeit"). For each specified day, the values "SECCI", "Temperatur", "Salinität", "NO2", "NO3", and "NOx" have to be predicted.

Note: Missing values or "NA" do not have to be predicted. Only valid values, indicated by the days given in the "bbdc_2023_AWI_data_evaluate_skeleton.csv" file, must be predicted.

Note: The data says "SECCI", but means "SECCHI".

For submission, the filled "bbdc_2023_AWI_data_evaluate_skeleton.csv" file must be submitted via the Submission System.

The evaluation metric is the Root Mean Squared Error (RMSE). Due to the variable scaling in the raw format, the input and reference data will be Z-normalized and the RMSE will be calculated on them.

Good luck and have fun!